Text.Documentalist.Types.DocBlock

Documentation

data DocBlock

A documentation block for a declaration.

Constructors

DocBlock

Fields

summary :: Paragraph

description :: [Paragraph]

parameters :: [DocParam]

example :: Maybe Code

result :: Maybe Result

data Paragraph

A newline-delimited section of text.

Constructors

TextParagraph [Span]

CodeBlock Code

QuotedText Paragraph

data Span

Represents a portion of text in a documentation string.

Constructors

PlainText String

Reference (Declaration (Maybe DocBlock))

WebLink String

InlineImage String

InlineCode Code

EmphasizedText Span

StrongText Span

UnderlinedText Span

newtype Code

A block or span of code in the source language.

Constructors

Code String

data DocParam

One of the parameters to a Function, or one of the values in an Enumeration.

Constructors

DocParam (Declaration (Maybe DocBlock)) [Span]

newtype Result

Describes the value that a Function returns to its caller.

Constructors

Result [Span]