Text.Documentalist.Types.Package

Documentation

newtype Identifier

An identifier, as it would be written in the source language.

Constructors

Identifier String

newtype Type

A type, as it would be declared in the source language.

Constructors

Type String

type SuperTypes = [Type]

Represents a list of types (or just a single type) that another type derives from.

type ResultTypes = [Type]

Represents a list of types that will be returned from a function or method.

type UnderlyingType = Maybe Type

Represents the underlying type for a declaration, if the source language supports it and one was provided.

data Declaration t

Any kind of documentable declaration, associated with data of type t.

Constructors

DecLeaf t Identifier DecFLeaf

DecNode t Identifier DecFNode [Declaration t]

data DecFNode

A declaration family type for a declaration that may contain more declarations inside of it

Constructors

Class SuperTypes
A class declaration.

Interface SuperTypes
An abstract interface, or Objective-C protocol definition.

Mixin Type
A mixin, or Objective-C category declaration.

Enumeration UnderlyingType
An enumeration.

Function ResultTypes
A function.

ClassMethod ResultTypes
A class method or static member function.

InstanceMethod ResultTypes
An instance method or member function.

data DecFLeaf

A declaration family type that can not carry any additional declarations

Constructors

Property UnderlyingType
A property declaration in a class.

Constant UnderlyingType
A constant within an enumeration or class, or outside of any scope.

Parameter UnderlyingType
The parameter to a function or method.

TypeAlias Type
A redeclaration of a type under a different name.

data Module t

A single module in the source language.

The declaration list should contain any top-level declarations in order of appearance.

Constructors

Module String [Declaration t]

data Package t

A package to treat as a single unit for the purposes of documentation generation.

Constructors

Package String [Module t]