Text.Documentalist.Writer

Documentation

module Text.Documentalist.Types.DocBlock

module Text.Documentalist.Types.Package

class MonadIO w => Writer w where

Generates formatted output from DocBlocks.

Methods

write :: Package (Maybe DocBlock) -> w ()

Writes formatted documentation to a destination determined by the specific Writer used.

Any errors will be indicated with a thrown Exception.